Fix stale user-string unable to login issue
This was fixed by checking if the user exists and, if they don't,
clearing the `user_string` cookie. Since it occurs in a `window.fetch`,
the best we can do is have the request fail, which should hopefully
prompt the user to try again. I will try to find a way to make the main
page reload.

Side note: also converted from two separate `await`s to one `await
Promise.all` for concurrency.